+++ backend/wikidata_utils.py
@@ -1,7 +1,7 @@
 from SPARQLWrapper import SPARQLWrapper, JSON
 from wikidata_models import WikidataEntity
 
-wikidata_label_query_cache = {}
+wikidata_label_query_cache = {} 
 
 def query_wikidata_for_label_and_description(items, sparql_endpoint):
     items = ' wd:'.join(items)
@@ -35,7 +35,7 @@ def query_wikidata_for_label_and_description(items, sparql_endpoint):
 
 def get_labels_and_descriptions(items, sparql_endpoint):
     response = dict()
-    missing_items = []
+    missing_items = {}
     for item in items:
         wp = WikidataEntity.query.filter_by(wd_id=item).first()
         if wp:
@@ -46,11 +46,15 @@ def get_labels_and_descriptions(items, sparql_endpoint):
                     desc = wp.description
                 response[item] = dict(label=label, description=desc)
             else:
-                missing_items.append(item)
+                if item not in wikidata_label_query_cache:
+                    missing_items[item]=True
         else:
-            missing_items.append(item)
+            if item not in wikidata_label_query_cache:
+                missing_items[item]=True
     try:
         if missing_items:
+            wikidata_label_query_cache.update(missing_items)
+            missing_items=list(missing_items.keys())
             additional_items = query_wikidata_for_label_and_description(
                 missing_items, sparql_endpoint)
             response.update(additional_items)