+++ electron/src/renderer/project/table-viewer/table-viewer.tsx
@@ -10,7 +10,7 @@ import 'ag-grid-community/dist/styles/ag-theme-balham.css';
 import { ChangeDetectionStrategyType } from 'ag-grid-react/lib/changeDetectionService';
 
 // console.log
-import { LOG, gridApiInterface, ErrorMessage, Cell } from '../../common/general';
+import { LOG, gridApiInterface, ErrorMessage, Cell, t2wmlColors } from '../../common/general';
 import RequestService, { IStateWithError } from '../../common/service';
 import ToastMessage from '../../common/toast';
 
@@ -175,22 +175,28 @@ class TableViewer extends Component<{}, TableState> {
 
 
   async handleSelectSheet(event: any) {
+    const sheetName = event.target.innerHTML;
     this.setState({
       showTable: false
     });
+    
+    // save prev yaml
+    await wikiStore.yaml.saveYaml();
+
     // remove current status
     this.updateSelectedCell(new Cell());
-    wikiStore.yaml.yamlContent = undefined;
+    wikiStore.yaml.yamlContent = '';
     wikiStore.output.isDownloadDisabled = true;
 
     // before sending request
     wikiStore.table.showSpinner = true;
     wikiStore.wikifier.showSpinner = true;
+    wikiStore.yaml.showSpinner = true;
 
     // send request
-    const sheetName = event.target.innerHTML;
     console.log("<TableViewer> -> %c/change_sheet%c for sheet: %c" + sheetName, LOG.link, LOG.default, LOG.highlight);
     try {
+
       await this.requestService.call(this, () => this.requestService.changeSheet(wikiStore.projects.current!.folder, sheetName));
       console.log("<TableViewer> <- %c/change_sheet%c with:", LOG.link, LOG.default);
 
@@ -207,6 +213,7 @@ class TableViewer extends Component<{}, TableState> {
     } finally {
       wikiStore.table.showSpinner = false;
       wikiStore.wikifier.showSpinner = false;
+      wikiStore.yaml.showSpinner = false;
     }
 
   }
@@ -339,7 +346,6 @@ class TableViewer extends Component<{}, TableState> {
       rowData2[row]["styles"][col] = Object.assign({}, rowData2[row]["styles"][col], style); // combine old and new styles
     }
 
-    console.log("setting state from updateStyleByArray", rowData2);
     this.setState({
       rowData: rowData2
     });
@@ -356,17 +362,10 @@ class TableViewer extends Component<{}, TableState> {
 
     let allIndices = Array<CellIndex>();
     const types = wikiStore.layers.type;
-    const typeStyles = new Map<string, any>([
-      ["data", { backgroundColor: "hsl(150, 50%, 90%)" }],
-      ["item", { backgroundColor: "hsl(200, 50%, 90%)" }],
-      ["qualifier", { backgroundColor: "hsl(250, 50%, 90%)" }],
-      ["reference", { backgroundColor: "hsl(150, 50%, 90%)" }],
-      ["minorError", { backgroundColor: '#FF8000' }],
-      ["majorError", { backgroundColor: '#FF3333' }]
-    ])
+
     for (const entry of types.entries) {
       allIndices = allIndices.concat(entry.indices)
-      const style = typeStyles.get(entry.type)
+      const style = utils.typeStyles.get(entry.type)
       if (style == undefined) {
         continue;
         //for now, pass. 
@@ -417,7 +416,6 @@ class TableViewer extends Component<{}, TableState> {
     }
 
 
-    console.log("setting state from toggle data", rowData2)
     this.setState({
       rowData: rowData2
     });
@@ -483,10 +481,12 @@ class TableViewer extends Component<{}, TableState> {
     if (wikiStore.projects.projectDTO) {
       const project = wikiStore.projects.projectDTO;
       const filename = project._saved_state.current_data_file
-      let sheetNames;
-      sheetNames = project.data_files[filename].val_arr
-      if (sheetNames == undefined) {
-        sheetNames = null;
+      let sheetNames = null;
+      if (filename) {
+        sheetNames = project.data_files[filename].val_arr;
+        if (sheetNames == undefined) {
+          sheetNames = null;
+        }
       }
 
       let multipleSheets = false;
@@ -514,7 +514,9 @@ class TableViewer extends Component<{}, TableState> {
         columnDefs: columns,
         rowData: rows,
         showTable: true,
-
+        styledCellsQnode: new Array<CellIndex>(),
+        styledCellsType: new Array<CellIndex>(),
+        styledCellsClean: new Array<CleanEntry>(),
       });
     }
     else {
@@ -526,6 +528,9 @@ class TableViewer extends Component<{}, TableState> {
         columnDefs: defaultColumns,
         rowData: defaultRows,
         showTable: true,
+        styledCellsQnode: new Array<CellIndex>(),
+        styledCellsType: new Array<CellIndex>(),
+        styledCellsClean: new Array<CleanEntry>(),
 
       });
     }
@@ -571,7 +576,7 @@ class TableViewer extends Component<{}, TableState> {
         <Card className="w-100 h-100 shadow-sm">
 
           {/* header */}
-          <Card.Header style={{ height: "40px", padding: "0.5rem 1rem", background: "#339966" }}>
+          <Card.Header style={{ height: "40px", padding: "0.5rem 1rem", background: t2wmlColors.TABLE }}>
 
             {/* title */}
             <div