+++ electron/src/renderer/project/wikifier/wikifier.tsx
@@ -4,7 +4,7 @@ import React, { Component, Fragment } from 'react';
 import { Button, Card, OverlayTrigger, Spinner, Tooltip } from 'react-bootstrap';
 
 // console.log
-import { LOG, ErrorMessage } from '../../common/general';
+import { LOG, ErrorMessage, t2wmlColors } from '../../common/general';
 import * as utils from '../../common/utils'
 
 import RequestService, { IStateWithError } from '../../common/service';
@@ -48,7 +48,7 @@ class Wikifier extends Component<WikifierProperties, WikifierState> {
       showSpinner: wikiStore.wikifier.showSpinner, //false,
 
       // wikifier data (from backend)
-      rowData: [], // e.g. [{ "context": "country", "col": "A", "row": "1", "value": "Burundi", "item": "Q967", "label": "Burundi", "description": "country in Africa" }]
+      rowData: [], // e.g. [{ "context": "country", "col": "A", "row": "1", "value": "Burundi", "subject": "Q967", "label": "Burundi", "description": "country in Africa" }]
 
       // call wikifier service
       showCallWikifier: false,
@@ -234,7 +234,7 @@ class Wikifier extends Component<WikifierProperties, WikifierState> {
 
           {/* header */}
           <Card.Header
-            style={{ height: "40px", padding: "0.5rem 1rem", background: "#006699" }}
+            style={{ height: "40px", padding: "0.5rem 1rem", background: t2wmlColors.WIKIFIER }}
             onClick={() => wikiStore.editors.nowShowing = "Wikifier"}
           >