+++ electron/src/renderer/project/yaml-editor.tsx
@@ -5,17 +5,19 @@ import MonacoEditor from 'react-monaco-editor';
 import yaml from 'js-yaml';
 
 // App
-import { Button, Card, OverlayTrigger, Tooltip } from 'react-bootstrap';
+import { Button, Card, OverlayTrigger, Spinner, Tooltip } from 'react-bootstrap';
 
 // console.log
-import { LOG, ErrorMessage } from '../common/general';
+import { LOG, ErrorMessage, t2wmlColors } from '../common/general';
 import RequestService, { IStateWithError } from '../common/service';
 import ToastMessage from '../common/toast';
 
 import { observer } from "mobx-react"
-import wikiStore from '../data/store';
+import wikiStore, { LayerState } from '../data/store';
 import { defaultYamlContent } from "./default-values";
 import { IReactionDisposer, reaction } from 'mobx';
+import SheetSelector from './table-viewer/sheet-selector';
+import { ProjectDTO } from '../common/dtos';
 
 
 interface yamlProperties {
@@ -23,13 +25,16 @@ interface yamlProperties {
 }
 
 interface yamlState extends IStateWithError {
-  yamlContent: string;
-  yamlTitle: string;
   yamlJson: JSON | null;
   isValidYaml: boolean;
   yamlParseErrMsg: string | null;
   yamlParseErrFloatMessage: string;
+  isYamlContentChanged: boolean;
+  isImportFile: boolean;
+  disableYaml: boolean;
+  isAddedYaml: boolean;
 }
+
 
 @observer
 class YamlEditor extends Component<yamlProperties, yamlState> {
@@ -45,22 +50,29 @@ class YamlEditor extends Component<yamlProperties, yamlState> {
     // init state
     this.state = {
       // yaml
-      yamlContent: defaultYamlContent,
-      yamlTitle: "",
       yamlJson: null,
       isValidYaml: true,
       yamlParseErrMsg: "",
       yamlParseErrFloatMessage: "",
       errorMessage: {} as ErrorMessage,
+      isYamlContentChanged: false,
+      isImportFile: false,
+      disableYaml: false,
+      isAddedYaml: false,
     };
 
     // init functions
     this.handleOpenYamlFile = this.handleOpenYamlFile.bind(this);
+    this.handleChangeFile = this.handleChangeFile.bind(this);
   }
 
   componentDidMount() {
+    if (!wikiStore.projects.projectDTO || !wikiStore.projects.projectDTO._saved_state.current_data_file) {
+      this.setState({disableYaml: true});
+    }
     this.disposeReaction = reaction(() => wikiStore.yaml.yamlContent, (newYamlContent) => this.updateYamlContent(newYamlContent));
     this.disposeReaction = reaction(() => wikiStore.yaml.yamlError, () => this.updateErrorFromStore());
+    this.disposeReaction = reaction(() => wikiStore.projects.projectDTO, (project) => { if (project) { this.updateYamlFiles(project); }});
   }
 
   componentWillUnmount() {
@@ -74,16 +86,19 @@ class YamlEditor extends Component<yamlProperties, yamlState> {
 
     // before sending request
     wikiStore.table.showSpinner = true;
+    wikiStore.yaml.showSpinner = true;
 
     // send request
     console.log("<YamlEditor> -> %c/upload_yaml%c for yaml regions", LOG.link, LOG.default);
     const formData = new FormData();
-    formData.append("yaml", this.state.yamlContent);
-    formData.append("title", this.state.yamlTitle);
+    formData.append("yaml", wikiStore.yaml.yamlContent);
+    formData.append("title", wikiStore.yaml.yamlName as string);
+    formData.append("sheetName", wikiStore.projects.projectDTO!._saved_state.current_sheet);
+
 
     try {
       await this.requestService.call(this, () => this.requestService.uploadYaml(wikiStore.projects.current!.folder, formData));
-      console.debug('Uploading yaml ', this.state.yamlContent);
+      console.debug('Uploading yaml ', wikiStore.yaml.yamlContent);
       console.log("<YamlEditor> <- %c/upload_yaml%c with:", LOG.link, LOG.default);
 
       // follow-ups (success)
@@ -93,17 +108,19 @@ class YamlEditor extends Component<yamlProperties, yamlState> {
     } catch {
     } finally {
       wikiStore.table.showSpinner = false;
+      wikiStore.yaml.showSpinner = false;
       wikiStore.table.isCellSelectable = true;
     }
 
   }
 
-  handleChangeYaml() {
+  handleEditYaml() {
     wikiStore.table.isCellSelectable = false;
 
-    // Talya: find out what's the right way to do this
     const yamlContent = (this.monacoRef.current as any).editor.getModel().getValue();
-    this.setState({ yamlContent: yamlContent });
+    wikiStore.yaml.yamlContent = yamlContent;
+    wikiStore.yaml.yamlhasChanged = true;
+    // this.setState({ isYamlContentChanged: true });
     try {
       const yamlJson = (yaml.safeLoad(yamlContent) as JSON);
       this.setState({
@@ -120,30 +137,45 @@ class YamlEditor extends Component<yamlProperties, yamlState> {
         yamlParseErrFloatMessage: err.stack,
       });
     }
+    if (yamlContent === defaultYamlContent) {
+      this.setState({ yamlParseErrMsg: '' });
+    }
   }
 
   handleOpenYamlFile(event: any) {
     // get file
     const file = (event.target as any).files[0];
     if (!file) return;
-    console.log("<YamlEditor> opened file: " + file.name);
+    let yamlName = file.name;
+    if (file.path.startsWith(wikiStore.projects.current!.folder)) {
+      yamlName = file.path.split(wikiStore.projects.current!.folder)[1].substring(1);
+    }
+    console.log("<YamlEditor> opened file: " + yamlName);
+    wikiStore.yaml.yamlName = yamlName;
+    wikiStore.yaml.yamlList = [...wikiStore.yaml.yamlList, yamlName];
 
     wikiStore.table.isCellSelectable = false;
 
     // upload local yaml
     const reader = new FileReader();
     reader.readAsText(file);
-    reader.onloadend = (() => {
-      const yamlContent = reader.result;
-      this.setState({ yamlContent: yamlContent as string, yamlTitle: file.name });
+    reader.onloadend = (async() => {
+      const yamlContent = reader.result as string;
+      wikiStore.yaml.yamlContent = yamlContent;
+      wikiStore.yaml.yamlName = yamlName;
+      wikiStore.yaml.yamlhasChanged = true;
       try {
         const yamlJson = (yaml.safeLoad((yamlContent as string))) as JSON;
         this.setState({
           yamlJson: yamlJson,
           isValidYaml: true,
           yamlParseErrMsg: null,
-          yamlParseErrFloatMessage: ''
+          yamlParseErrFloatMessage: '',
+          isImportFile: true
         });
+
+        // Save yaml when importing yaml
+        await wikiStore.yaml.saveYaml();
       } catch (err) {
         this.setState({
           yamlJson: null,
@@ -152,12 +184,15 @@ class YamlEditor extends Component<yamlProperties, yamlState> {
           yamlParseErrFloatMessage: err.stack
         });
       }
+      if (yamlContent === defaultYamlContent) {
+        this.setState({ yamlParseErrMsg: '' });
+      }
     });
   }
 
   updateYamlContent(yamlContent: string | undefined) {
     const newYamlContent = yamlContent || defaultYamlContent;
-    this.setState({ yamlContent: newYamlContent });
+    wikiStore.yaml.yamlContent = newYamlContent;
     try {
       const yamlJson = (yaml.safeLoad(newYamlContent)) as JSON;
       this.setState({
@@ -174,24 +209,143 @@ class YamlEditor extends Component<yamlProperties, yamlState> {
         yamlParseErrFloatMessage: err.stack
       });
     }
+    if (yamlContent === defaultYamlContent) {
+      this.setState({ yamlParseErrMsg: '' });
+    }
   }
 
   updateErrorFromStore(){
     const yamlError = wikiStore.yaml.yamlError;
-    if (yamlError && yamlError!=""){
+    if (yamlError && yamlError!="") {
       console.log("Errors while applying yaml:")
       console.log(yamlError);
       console.log(wikiStore.layers.error);
-      this.setState({
-        yamlParseErrMsg: "⚠️ " + yamlError
-      });
+      if (wikiStore.yaml.yamlContent !== defaultYamlContent) {
+        this.setState({
+          yamlParseErrMsg: "⚠️ " + yamlError
+        });
+      }
+    }
+  }
+
+  updateYamlFiles(project: ProjectDTO) {
+    if (wikiStore.projects.projectDTO && wikiStore.projects.projectDTO._saved_state.current_data_file) {
+      this.setState({disableYaml: false});
     }
 
+    const dataFile = project._saved_state.current_data_file;
+    const sheetName = project._saved_state.current_sheet;
+    if (dataFile) {
+      if (project!.yaml_sheet_associations[dataFile] && project!.yaml_sheet_associations[dataFile][sheetName]) {
+        wikiStore.yaml.yamlList = project!.yaml_sheet_associations[dataFile][sheetName].val_arr;
+        wikiStore.yaml.yamlName = project!.yaml_sheet_associations[dataFile][sheetName].selected;
+      } else {
+        // this.setState({isAddedYaml: true});
+        let yamlToCurrentSheet = project?._saved_state.current_sheet;
+        if (yamlToCurrentSheet.endsWith('.csv')) {
+          yamlToCurrentSheet = yamlToCurrentSheet.split('.csv')[0];
+        } 
+        if (!yamlToCurrentSheet.endsWith('.yaml')) {
+          yamlToCurrentSheet += '.yaml';
+        }
 
+        wikiStore.yaml.yamlName = yamlToCurrentSheet;
+        wikiStore.yaml.yamlList = [yamlToCurrentSheet];
+      }
+    } else {
+      wikiStore.yaml.yamlName = '';
+      wikiStore.yaml.yamlList = [];
+    }
+  }
+
+  async handleChangeFile(event: any) {
+    const yaml = event.target.innerHTML;
+    
+    // save prev yaml
+    await wikiStore.yaml.saveYaml();
+
+    wikiStore.yaml.yamlName = yaml;
+
+    wikiStore.output.isDownloadDisabled = true;
+
+    // before sending request
+    wikiStore.table.showSpinner = true;
+    wikiStore.wikifier.showSpinner = true;
+    wikiStore.yaml.showSpinner = true;
+
+    // send request
+    try {
+      await this.requestService.call(this, () => this.requestService.changeYaml(wikiStore.projects.current!.folder, yaml));
+
+      if (wikiStore.yaml.yamlContent) {
+        wikiStore.table.isCellSelectable = true;
+        wikiStore.output.isDownloadDisabled = false;
+      } else {
+        wikiStore.table.isCellSelectable = false;
+      }
+    } catch {
+    } finally {
+      wikiStore.table.showSpinner = false;
+      wikiStore.wikifier.showSpinner = false;
+      wikiStore.yaml.showSpinner = false;
+    }
+  }
+
+  async renameYaml(val: string, index: number) {
+   // before sending request
+    wikiStore.table.showSpinner = true;
+    wikiStore.yaml.showSpinner = true;
+    const oldName = wikiStore.yaml.yamlList[index];
+
+    // Check if this yaml file exist, if not- save it before.
+    if (!wikiStore.projects.projectDTO?.yaml_files.includes(oldName)) {
+      await wikiStore.yaml.saveYaml();
+    }
+
+    // send request
+    const formData = new FormData();
+    formData.append("old_name", oldName);
+    formData.append("new_name", val);
+
+    try {
+      await this.requestService.call(this, () => this.requestService.renameYaml(wikiStore.projects.current!.folder, formData));
+      // update yaml files according to received project.
+      wikiStore.yaml.yamlList = wikiStore.projects.projectDTO!.yaml_files;
+    } catch {
+      console.error("Rename yaml failed.");
+    } finally {
+      wikiStore.table.showSpinner = false;
+      wikiStore.yaml.showSpinner = false;
+    }
+  }
+
+  async addYaml() {
+    await wikiStore.yaml.saveYaml();
+
+    this.setState({
+      isAddedYaml: true,
+    });
+    let i = 1;
+    let sheetName = wikiStore.projects.projectDTO!._saved_state.current_sheet;
+    // remove .csv from sheet name
+    if (sheetName.endsWith('.csv')) {
+      sheetName = sheetName.split('.csv')[0];
+    }
+    let yamlName = sheetName + "-" + i + ".yaml";  
+    while (wikiStore.projects.projectDTO!.yaml_files.includes(yamlName)) {
+      i++;
+      yamlName = sheetName + "-" + i + ".yaml"; 
+    }
+    
+    wikiStore.yaml.yamlContent = defaultYamlContent;
+    wikiStore.yaml.yamlName = yamlName;
+    wikiStore.yaml.yamlList.push(yamlName);
+
+    wikiStore.layers = new LayerState();
   }
 
   render() {
-    const { yamlContent } = this.state;
+    const yamlContent = wikiStore.yaml.yamlContent;
 
     // render upload tooltip
     const uploadToolTipHtml = (
@@ -213,7 +367,7 @@ class YamlEditor extends Component<yamlProperties, yamlState> {
 
           {/* header */}
           <Card.Header
-            style={{ height: "40px", padding: "0.5rem 1rem", background: "#006699" }}
+            style={{ height: "40px", padding: "0.5rem 1rem", background: t2wmlColors.YAML }}
             onClick={() => wikiStore.editors.nowShowing = "YamlEditor"}
           >
 
@@ -244,17 +398,24 @@ class YamlEditor extends Component<yamlProperties, yamlState> {
               id="file_yaml"
               accept=".yaml"
               style={{ display: "none" }}
+              disabled={this.state.disableYaml || wikiStore.yaml.showSpinner}
               onChange={this.handleOpenYamlFile}
               onClick={(event) => { (event.target as HTMLInputElement).value = '' }}
             />
 
           </Card.Header>
 
+          {/* loading spinner */}
+          <div className="mySpinner" hidden={!wikiStore.yaml.showSpinner}>
+              <Spinner animation="border" />
+          </div>
+
           {/* yaml editor */}
           <Card.Body
             className="w-100 h-100 p-0"
             style={(this.props.isShowing) ? { overflow: "hidden" } : { display: "none" }}
           >
+            { ! this.state.disableYaml ?
             <MonacoEditor ref={this.monacoRef}
               width="100%"
               height="100%"
@@ -278,20 +439,20 @@ class YamlEditor extends Component<yamlProperties, yamlState> {
                 },
                 showFoldingControls: 'always',
               }}
-              onChange={() => this.handleChangeYaml()}
+              onChange={() => this.handleEditYaml()}
               editorDidMount={(editor) => {
                 editor.getModel()?.updateOptions({ tabSize: 2 });
                 setInterval(() => { editor.layout(); }, 200);  // automaticLayout above misses trigger opening, so we've replaced it
                 // This is better done by catching the trigger event and calling editor.layout there,
                 // once we figure out how to catch the trigger event.
               }}
-            />
+            /> : null }
           </Card.Body>
 
           {/* card footer */}
           <Card.Footer
             style={
-              (this.props.isShowing) ? { height: "40px", padding: "0.5rem 1rem", background: "whitesmoke" } : { display: "none" }
+              (this.props.isShowing) ? { background: "whitesmoke" } : { display: "none" }
             }
           >
 
@@ -301,8 +462,8 @@ class YamlEditor extends Component<yamlProperties, yamlState> {
               style={{
                 // fontFamily: "Menlo, Monaco, \"Courier New\", monospace",
                 fontSize: "14px",
-                color: "#990000",
-                width: "calc(100% - 60px)",
+                color: t2wmlColors.OUTPUT,
+                width: "calc(100% - 10px)",
                 cursor: "help"
               }}
               title={this.state.yamlParseErrFloatMessage}
@@ -314,12 +475,35 @@ class YamlEditor extends Component<yamlProperties, yamlState> {
             <Button
               className="d-inline-block float-right"
               size="sm"
-              style={{ borderColor: "#006699", background: "#006699", padding: "0rem 0.5rem" }}
+              style={{ borderColor: t2wmlColors.YAML, background: t2wmlColors.YAML, padding: "0rem 0.5rem" }}
               onClick={() => this.handleApplyYaml()}
-              disabled={!this.state.isValidYaml}
+              disabled={!this.state.isValidYaml || this.state.disableYaml || wikiStore.yaml.showSpinner}
             >
               Apply
             </Button>
+            
+            {/* <div
+              id="yamlSelector" // apply custom scroll bar
+              style={{
+                height: "55px",
+                padding: "0.5rem 0.75rem",
+                background: "whitesmoke",
+                // overflow: "scroll hidden", // safari does not support this
+                overflowX: "scroll",
+                overflowY: "hidden",
+                whiteSpace: "nowrap"
+              }}
+            > */}
+              <SheetSelector
+                sheetNames={wikiStore.yaml.yamlList}
+                currSheetName={wikiStore.yaml.yamlName}
+                itemType="file"
+                handleSelectSheet={(event) => this.handleChangeFile(event)}
+                handleAddItem={() => this.addYaml()}
+                disableAdd={wikiStore.yaml.yamlContent === defaultYamlContent}
+                handleDoubleClickItem={(val, index) => this.renameYaml(val, index)}
+              />
+            {/* </div> */}
           </Card.Footer>
         </Card>
       </Fragment>